Refactor: character 도멤메인 관련 수정 및 새로운 푸키 ë°발급ê 기능
- UserCharacter필드에 낙관적 락을 위한 version 필드 추가

+++ backend/src/main/java/com/ssafy/pookie/character/dto/RepCharacterResponseDto.java
@@ -11,7 +11,7 @@ import lombok.*;
 @Builder
 public class RepCharacterResponseDto {
 
-    private int userCharacterId;
+    private Long userCharacterId;
     private String characterName;
     private int step;
     private int exp;
